requirements.txt
requests
openai
python-dotenv
numpy
#cognee==0.5.0
git+https://github.com/tiagoriosrocha/cognee.git@adapters-0.5.0 #utilizo o branch adapters-0.5.0 do cognee, pois a versão 0.5.0 não tem suporte para adapters
protego>=0.1 
playwright>=1.9.0 
transformers
python-certifi-win32